refactor(ffi): delete the output DTO; retain ffi-wasi slimmed (#1419 Stage C)
Final stage of Phase 5. After Stage B the component converts core->WIT directly,
so rustledger-ffi-wasi's Directive->JSON output DTO is dead. Remove it.

- Moved compute_directive_hash (core->hash, DTO-free; the component uses it) into
  a new src/hash.rs before deleting convert.rs.
- Deleted src/convert.rs (directive_to_json, value_to_json, position_to_json,
  value_datatype) and the now-unused directive-output DTO types from
  types/output.rs (DirectiveJson, Posting, Amount, Meta, CostNumber, TypedValue,
  PostingCost, meta_value_to_json + their wire-shape tests). Kept Error,
  LedgerOptions, Plugin, Include — helpers' LoadResult/FileLoad return these.
- Updated lib.rs / types/mod.rs re-exports; reworded the component's doc comments
  that named the deleted convert:: paths (keeps cargo doc -D warnings clean).

Crate-fate decision (the open item in #1419): **retain rustledger-ffi-wasi as the
slimmed FFI-support helpers crate** (loader orchestration + WIT-input construction
+ compute_directive_hash). It is deliberately NOT folded into rustledger-loader:
its survivors are FFI glue — an 864-line WIT-input parser and Serialize-shaped
duplicates of loader's native Options/errors — which belong out of the core loader
crate. Documented in CLAUDE.md, architecture.md, and both crate READMEs.

Phase 5 is complete: the wasip1 JSON-RPC surface (Stage A), the DTO middle layer
(Stage B/C), and the wire-format tests are all gone; the component is the sole,
typed, core->WIT embedding surface.

Verified: native workspace build + clippy clean, cargo doc -D warnings clean,
15/15 wasmtime-host parity tests pass on the real wasip2 component, fmt clean.

Closes #1419
