+++ crates/rustledger-ffi-wasi/src/types/mod.rs
@@ -4,7 +4,4 @@ pub mod input;
 pub mod output;
 
 pub use input::input_entry_to_directive;
-pub use output::{
-    Amount, CostNumber, DirectiveJson, Error, Include, LedgerOptions, Meta, Plugin, Posting,
-    PostingCost, TypedValue, meta_value_to_json,
-};
+pub use output::{Error, Include, LedgerOptions, Plugin};
