+++ crates/rustledger-ffi-wasi/src/types/output.rs
@@ -2,42 +2,7 @@
 
 use std::collections::HashMap;
 
-use rustledger_core::{MetaValue, Metadata};
 use serde::Serialize;
-
-/// Metadata includes filename, lineno, hash, plus any user-defined key-value pairs.
-#[derive(Serialize, Default)]
-pub struct Meta {
-    pub filename: String,
-    pub lineno: u32,
-    /// Entry hash (SHA256 of canonical representation).
-    pub hash: String,
-    #[serde(flatten)]
-    pub user: HashMap<String, serde_json::Value>,
-}
-
-impl Meta {
-    pub fn new(filename: &str, lineno: u32, hash: String, directive_meta: &Metadata) -> Self {
-        let mut user = HashMap::new();
-        for (key, value) in directive_meta {
-            user.insert(key.clone(), meta_value_to_json(value));
-        }
-        Self {
-            filename: filename.to_string(),
-            lineno,
-            hash,
-            user,
-        }
-    }
-}
-
-/// Convert `MetaValue` to its canonical JSON form.
-///
-/// Thin wrapper over the single source [`rustledger_core::meta_value_to_json`];
-/// kept so call sites and the `types` module API read in ffi-wasi terms.
-pub fn meta_value_to_json(value: &MetaValue) -> serde_json::Value {
-    rustledger_core::meta_value_to_json(value)
-}
 
 #[derive(Serialize, Clone)]
 pub struct Error {
@@ -84,181 +49,6 @@ impl Error {
         self.phase = "validate".to_string();
         self
     }
-}
-
-#[derive(Serialize, Clone)]
-pub struct Amount {
-    pub number: String,
-    pub currency: String,
-}
-
-/// Wire-format of the numeric component of a [`PostingCost`].
-///
-/// Mirrors the host's `rustledger_core::CostNumber` enum so JSON
-/// consumers see the same mutual exclusion the host enforces. The
-/// `kind` tag is the discriminator; consumers should switch on it
-/// rather than probing for present-but-null fields.
-#[derive(Serialize)]
-#[serde(tag = "kind", rename_all = "snake_case")]
-pub enum CostNumber {
-    /// Per-unit cost (e.g., `{100 USD}`).
-    PerUnit {
-        /// Per-unit value.
-        value: String,
-    },
-    /// Total cost as written (e.g., `{{1000 USD}}`), pre-booking.
-    Total {
-        /// Total value.
-        value: String,
-    },
-    /// Post-booking: derived per-unit plus preserved source total.
-    PerUnitFromTotal {
-        /// Derived per-unit.
-        per_unit: String,
-        /// Original total.
-        total: String,
-    },
-}
-
-#[derive(Serialize)]
-pub struct PostingCost {
-    /// Cost number (per-unit, total, or post-booking pair).
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub number: Option<CostNumber>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub currency: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub date: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub label: Option<String>,
-}
-
-/// A typed value preserving the original type from the beancount source.
-#[derive(Serialize)]
-#[serde(rename_all = "lowercase")]
-pub struct TypedValue {
-    #[serde(rename = "type")]
-    pub value_type: &'static str,
-    pub value: serde_json::Value,
-}
-
-impl TypedValue {
-    pub fn from_meta_value(mv: &MetaValue) -> Self {
-        // The `type` tag and the JSON `value` are both single-sourced in core,
-        // so the tagged form can never drift from the plain `meta` map form.
-        Self {
-            value_type: rustledger_core::meta_value_type_tag(mv),
-            value: rustledger_core::meta_value_to_json(mv),
-        }
-    }
-}
-
-#[derive(Serialize)]
-pub struct Posting {
-    pub account: String,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub units: Option<Amount>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub cost: Option<PostingCost>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub price: Option<Amount>,
-    /// Posting-level flag (e.g., "!" for pending).
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub flag: Option<String>,
-    #[serde(skip_serializing_if = "HashMap::is_empty")]
-    pub meta: HashMap<String, serde_json::Value>,
-}
-
-#[derive(Serialize)]
-#[serde(tag = "type", rename_all = "snake_case")]
-pub enum DirectiveJson {
-    Transaction {
-        date: String,
-        flag: String,
-        #[serde(skip_serializing_if = "Option::is_none")]
-        payee: Option<String>,
-        #[serde(skip_serializing_if = "Option::is_none")]
-        narration: Option<String>,
-        tags: Vec<String>,
-        links: Vec<String>,
-        postings: Vec<Posting>,
-        meta: Meta,
-    },
-    Open {
-        date: String,
-        account: String,
-        currencies: Vec<String>,
-        #[serde(skip_serializing_if = "Option::is_none")]
-        booking: Option<String>,
-        meta: Meta,
-    },
-    Close {
-        date: String,
-        account: String,
-        meta: Meta,
-    },
-    Balance {
-        date: String,
-        account: String,
-        amount: Amount,
-        /// Explicit tolerance (e.g., "~ 0.01").
-        #[serde(skip_serializing_if = "Option::is_none")]
-        tolerance: Option<String>,
-        meta: Meta,
-    },
-    Pad {
-        date: String,
-        account: String,
-        source_account: String,
-        meta: Meta,
-    },
-    Commodity {
-        date: String,
-        currency: String,
-        meta: Meta,
-    },
-    Price {
-        date: String,
-        currency: String,
-        amount: Amount,
-        meta: Meta,
-    },
-    Event {
-        date: String,
-        event_type: String,
-        value: String,
-        meta: Meta,
-    },
-    Note {
-        date: String,
-        account: String,
-        comment: String,
-        meta: Meta,
-    },
-    Document {
-        date: String,
-        account: String,
-        path: String,
-        /// Tags attached to the document directive (issue #1144).
-        #[serde(skip_serializing_if = "Vec::is_empty")]
-        tags: Vec<String>,
-        /// Links attached to the document directive (issue #1144).
-        #[serde(skip_serializing_if = "Vec::is_empty")]
-        links: Vec<String>,
-        meta: Meta,
-    },
-    Query {
-        date: String,
-        name: String,
-        query_string: String,
-        meta: Meta,
-    },
-    Custom {
-        date: String,
-        custom_type: String,
-        values: Vec<TypedValue>,
-        meta: Meta,
-    },
 }
 
 #[derive(Serialize)]
@@ -354,67 +144,6 @@ pub struct Include {
 #[cfg(test)]
 mod tests {
     use super::*;
-
-    // ===== Cost-number wire-format tests (#1164) =====
-    //
-    // These pin the JSON shape FFI consumers depend on. Mirrors host
-    // `CostNumber` exactly — any silent shape change is a wire break.
-
-    #[test]
-    fn cost_number_per_unit_serializes_with_kind_tag() {
-        let cn = CostNumber::PerUnit {
-            value: "100".into(),
-        };
-        let json = serde_json::to_value(&cn).unwrap();
-        assert_eq!(
-            json,
-            serde_json::json!({"kind": "per_unit", "value": "100"})
-        );
-    }
-
-    #[test]
-    fn cost_number_total_serializes_with_kind_tag() {
-        let cn = CostNumber::Total {
-            value: "1500".into(),
-        };
-        let json = serde_json::to_value(&cn).unwrap();
-        assert_eq!(json, serde_json::json!({"kind": "total", "value": "1500"}));
-    }
-
-    #[test]
-    fn cost_number_per_unit_from_total_carries_both_values() {
-        let cn = CostNumber::PerUnitFromTotal {
-            per_unit: "150".into(),
-            total: "300".into(),
-        };
-        let json = serde_json::to_value(&cn).unwrap();
-        // Load-bearing: the preserved total survives serialization so
-        // downstream consumers don't have to redivide. This is what
-        // the pre-PR shape silently lost.
-        assert_eq!(
-            json,
-            serde_json::json!({
-                "kind": "per_unit_from_total",
-                "per_unit": "150",
-                "total": "300",
-            })
-        );
-    }
-
-    #[test]
-    fn posting_cost_omits_number_when_none() {
-        let pc = PostingCost {
-            number: None,
-            currency: Some("USD".into()),
-            date: None,
-            label: None,
-        };
-        let json = serde_json::to_string(&pc).unwrap();
-        // `skip_serializing_if = "Option::is_none"` keeps the JSON
-        // shape lean — a bare `{USD}` lot match has no `number` key.
-        assert!(!json.contains("number"));
-        assert!(json.contains("currency"));
-    }
 
     #[test]
     fn test_error_default_phase_is_parse() {
